&stats=0, hide the "Show Stats" option from the right-click menu
add an extended sub-option for the opt-in parameter &stats.
When &stats equals any value in ["false","0","no","off"], the "Show Stats" option from the right-click menu will be hidden. 

The &stats info is mainly used for debug, hide those stats info in Prod system might help to secure the system better. When necessary, the "Ctrl + Left-click" still can be used to call out the "stats menu" as a hidden short cut.